Get picture URLs from API response body
The default behaviour for `graph.facebook.com/me/picture` and other
related endpoints is to return a 302 redirect to the profile photo url.
To make these `.../picture` endpoints return responses similar to the
rest of the Graph API, we must add the `redirect=false` parameter [0].

Previously, we would make the request to the graph API without this
parameter, and then extract the photo URL from the 'Location' header in
the HTTP response. This is a little unconventional, and may be
responsible for issue #308, although I haven't tested the failure mode
in a batch query here. Instead, we can explicitly request the full API
response (via #get_user_picture_data), and then parse out the URL field.

[0] The current version of the Graph API documentation doesn't seem to
  note this particular wart; I've opened a bug against their
  documentation:

    https://developers.facebook.com/bugs/383635598495430/

  In the meantime, you can find an old version describing this
  behavior thanks to the Wayback Machine:

    http://web.archive.org/web/20131020151411/https://developers.facebook.com/docs/reference/api/using-pictures/