Match get_user_picture_data mocks to API results
This mock response didn't include the "data" wrapper around the result,
meaning that we weren't testing against the correct response.

The correctness of the format of this mock can be validated by visiting
any of the following urls:

  http://graph.facebook.com/lukeshephard/picture?redirect=false
  http://graph.facebook.com/v1.0/lukeshephard/picture?redirect=false
  http://graph.facebook.com/v2.0/100005142699145/picture?redirect=false
  http://graph.facebook.com/v2.3/100005142699145/picture?redirect=false

Note: In newer versions of the Graph API we can no longer directly query
  the username 'lukeshephard', so we must first dereference it to get the
  user's Facebook ID.

While updating the response here, I changed the picture url to match the
url response from `/lukeshephard/picture` instead of
`/lukeshephard/picture?type=large`. This should help make clear that the
get_user_picture_data response doesn't somehow infer that we should be
returning a large image.